* changed bucket addressing within a branch to use int instead of byte (underlying memory structures can still be bytes, but now they are unsigned) allowing for more than 127 buckets per branch (e.g. 169 in preflop)

+++ trunk/ao/ai/equilibrium/limit_cfr/CfrBot2.java
@@ -56,12 +56,12 @@ public class CfrBot2 extends AbstractPlayer
     private Round          prevRound;
 
     @SuppressWarnings({"FieldCanBeLocal"})
-    private byte relBucket;
+    private int  relBucket;
     private char roundBucket;
 
-    private byte holeBucket;
-    private byte flopBucket;
-    private byte turnBucket;
+    private int  holeBucket;
+    private int  flopBucket;
+    private int  turnBucket;
 
     private final List<String> handProbabilities =
             new ArrayList<String>();
@@ -236,7 +236,8 @@ public class CfrBot2 extends AbstractPlayer
 
         River canonRiver =
                 canonTurn.addRiver( cards.community().river() );
-        byte riverBucket = ABS.tree(true).getRiver(canonRiver.canonIndex());
+        int  riverBucket = ABS.tree(true)
+                .getRiver(canonRiver.canonIndex());
 
         roundBucket = ABS.decoder().decode(
                         holeBucket, flopBucket, turnBucket, riverBucket);
